Testing of code and windows::synchro
Added distant::process as an extension of windows::synchro. Refer to source documentation for design choice reasoning.

Removed storage value from memory::value. This was not needed.

Moved distant::handle into windows::handle, since this is implementation-specific.

Further work on memory::vm so that it works properly in all cases.